Redirect to the latest version of the docs.
Now that we have the `latest` symlink, redirect from the homepage to the latest
version of the docs rather than hardcoding a specific version.

The docs themselves will specify what version of JanusGraph they apply to, so
there won't be any user confusion, and they'll always have the latest version of
the docs.

Once we have several versions, we'll update the index page to instead list all
the available doc versions and let the user choose explicitly.

Signed-off-by: Misha Brukman <mbrukman@google.com>